requirements.txt
torch
torchvision
tensorflow
numpy
matplotlib
seaborn
scikit-learn
Pillow
h5py